Rename CrossOriginHelper to ...Internal; improve package-info
